requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
python-decouple>=3.8
dj-database-url>=2.1
Pillow>=10.0
psycopg2-binary>=2.9   # remove if using SQLite only
